Redirect automatically to the first organisation
