+++ app/routes/_dashboard+/dashboard.$organisationId.tsx
@@ -1,5 +1,6 @@
 import { invariantResponse } from '@epic-web/invariant'
-import { Box, Button, Stack, Typography } from '@mui/material'
+import { Box, Button, Grid, Stack, Typography } from '@mui/material'
+import { ListChecks as ListChecksIcon } from '@phosphor-icons/react/dist/ssr/ListChecks'
 import { Plus as PlusIcon } from '@phosphor-icons/react/dist/ssr/Plus'
 import {
 	type LoaderFunctionArgs,
@@ -7,6 +8,8 @@ import {
 	type MetaFunction,
 } from '@remix-run/node'
 import { useLoaderData } from '@remix-run/react'
+import { Summary } from '#app/components/dashboard/overview/summary.js'
+import { dailyTotalsApiResponseSchema } from '#app/types/bigmelo/dailyTotals.js'
 import { organisationsApiResponseSchema } from '#app/types/bigmelo/organisations.js'
 import { get } from '#app/utils/api.js'
 import { requireAuthedSession } from '#app/utils/auth.server.js'
@@ -16,17 +19,16 @@ import { verifyZodSchema } from '#app/utils/verifyZodSchema.js'
 export const meta: MetaFunction = () => [{ title: 'Organización' }]
 
 export async function loader({ request, params }: LoaderFunctionArgs) {
-	const organisationId = params.organisationId ?? ''
-	const parsedOrganisationId = parseInt(organisationId, 10)
-	if (isNaN(parsedOrganisationId)) {
-		invariantResponse(false, 'Organisation not found', {
-			status: 404,
-		})
-	}
-
 	const { authHeader } = await requireAuthedSession(request)
-
 	try {
+		const organisationId = params.organisationId ?? ''
+		const parsedOrganisationId = parseInt(organisationId, 10)
+		if (isNaN(parsedOrganisationId)) {
+			invariantResponse(false, 'Organisation not found', {
+				status: 404,
+			})
+		}
+
 		const organisationsResponse = await get('/v1/organization', {
 			headers: {
 				...authHeader,
@@ -46,7 +48,20 @@ export async function loader({ request, params }: LoaderFunctionArgs) {
 			status: 404,
 		})
 
+		// TODO: Implement the daily totals API per organisation
+		const dailyTotalsResponse = await get('/v1/admin-dashboard/daily-totals', {
+			headers: {
+				...authHeader,
+			},
+		})
+
+		const dailyTotals = verifyZodSchema(
+			dailyTotalsResponse.data,
+			dailyTotalsApiResponseSchema,
+		)
+
 		return json({
+			dailyTotals: dailyTotals.data,
 			organisation,
 		})
 	} catch (error) {
@@ -55,7 +70,7 @@ export async function loader({ request, params }: LoaderFunctionArgs) {
 }
 
 export default function Index() {
-	const { organisation } = useLoaderData<typeof loader>()
+	const { organisation, dailyTotals } = useLoaderData<typeof loader>()
 	return (
 		<Box
 			sx={{
@@ -79,6 +94,64 @@ export default function Index() {
 						Crear un Proyecto
 					</Button>
 				</Stack>
+				{dailyTotals && (
+					<Grid container spacing={4}>
+						<Grid item md={4} xs={12}>
+							<Summary
+								amount={dailyTotals.newLeads}
+								diff={0}
+								icon={ListChecksIcon}
+								title="New Leads"
+								trend="up"
+							/>
+						</Grid>
+						<Grid item md={4} xs={12}>
+							<Summary
+								amount={dailyTotals.newUsers}
+								diff={0}
+								icon={ListChecksIcon}
+								title="New Users"
+								trend="up"
+							/>
+						</Grid>
+						<Grid item md={4} xs={12}>
+							<Summary
+								amount={dailyTotals.newMessages}
+								diff={0}
+								icon={ListChecksIcon}
+								title="New Messages"
+								trend="up"
+							/>
+						</Grid>
+						<Grid item md={4} xs={12}>
+							<Summary
+								amount={dailyTotals.newWhatsappMessages}
+								diff={0}
+								icon={ListChecksIcon}
+								title="New WhatsApp Messages"
+								trend="up"
+							/>
+						</Grid>
+						<Grid item md={4} xs={12}>
+							<Summary
+								amount={dailyTotals.newAudioMessages}
+								diff={0}
+								icon={ListChecksIcon}
+								title="New Audio Messages"
+								trend="up"
+							/>
+						</Grid>
+						<Grid item md={4} xs={12}>
+							<Summary
+								amount={dailyTotals.dailyChats}
+								diff={0}
+								icon={ListChecksIcon}
+								title="Daily Chats"
+								trend="up"
+							/>
+						</Grid>
+					</Grid>
+				)}
 			</Stack>
 		</Box>
 	)
